Worked on Battle, and made some stuff neater
Battle class now works; added an accompanying test to show it off.
changed the constructors for items and moves so that they are more
similar to each other, and have willpower.
Changed all references of "speed" to "spd" so that it follows the
convention of abbreviating stats to three letters(changes made in
Character, Enemy, Player, and Battle).
Changed the test class to work with the new constructors for items and
moves.